+++ common/src/main/scala/collection/compat/ScalaNative.scala
@@ -38,7 +38,7 @@ final class ScalaNative[+A](val xs: Iterable[A], val counter: Counter) extends a
   def labeled(label: String): MapTo[A]               = this
   def reverse: MapTo[A]                              = xs.reverse
   def sized(size: Size): MapTo[A]                    = this
-  def slice(range: Interval): MapTo[A]               = xs.slice(range.start, range.end)
+  def slice(range: Interval): MapTo[A]               = xs.slice(range.start.toInt, range.end.toInt)
   def withFilter(p: Predicate[A]): MapTo[A]          = xs filter p
 
   def description = xs.shortClass
