+++ common/src/main/scala/core/PspInt.scala
@@ -2,7 +2,7 @@ package psp
 package core
 
 import java.{ lang => jl }
-import psp.collection.{ Direct, Foreach, IntRange }
+import psp.collection.{ Direct, Foreach, LongRange }
 
 final class PspInt(val self: Int) extends AnyVal with Ordered[Int] {
   def compare(rhs: Int): Int = jl.Integer.compare(self, rhs)
@@ -16,7 +16,7 @@ final class PspInt(val self: Int) extends AnyVal with Ordered[Int] {
   def hex: String    = jl.Integer.toHexString(self)
   def octal: String  = jl.Integer.toOctalString(self)
 
-  def until(end: Int): IntRange        = IntRange.until(self, end)
-  def to(end: Int): IntRange           = IntRange.to(self, end)
+  def until(end: Long): LongRange        = LongRange.until(self, end)
+  def to(end: Long): LongRange           = LongRange.to(self, end)
   def times[A](body: => A): Foreach[A] = Direct.fill(self)(body)
 }
